# api/oas/oas_schemas_gen.go
// Code generated by ogen, DO NOT EDIT.

package api

// OnboardForbidden is response for Onboard operation.
type OnboardForbidden struct{}

func (*OnboardForbidden) onboardRes() {}

// OnboardOK is response for Onboard operation.
type OnboardOK struct{}

func (*OnboardOK) onboardRes() {}

// OnboardUnauthorized is response for Onboard operation.
type OnboardUnauthorized struct{}

func (*OnboardUnauthorized) onboardRes() {}

// Specification on which namespace to create.
// Ref: #/components/schemas/OnboardingRequest
type OnboardingRequest struct {
	Group string `json:"group"`
}

// GetGroup returns the value of Group.
func (s *OnboardingRequest) GetGroup() string {
	return s.Group
}

// SetGroup sets the value of Group.
func (s *OnboardingRequest) SetGroup(val string) {
	s.Group = val
}